split fragment NavEventNavigationHandler to support plain NavEventNavigator
